Match menu_radio_codec_helper_8004158C

+++ src/Menu/menu_radio_8004D334.c
@@ -4,5 +4,5 @@ extern Actor_MenuMan gMenuMan_800BD360;
 
 int menu_radio_8004D334(GV_PAD *pPad)
 {
-    return menu_radio_do_file_mode_8004C418((int)&gMenuMan_800BD360, pPad);
+    return menu_radio_do_file_mode_8004C418(&gMenuMan_800BD360, pPad);
 }
